made schedules filter persistent and added clear selection button
